+++ src/main/java/cart/controller/customer/CustomerProductController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.customer;
 
 import cart.controller.dto.ProductDto;
 import cart.service.ProductService;
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import java.util.List;
 
 @Controller
-public class ShoppingController {
+public class CustomerProductController {
 
     private final ProductService productService;
 
-    public ShoppingController(final ProductService productService) {
+    public CustomerProductController(final ProductService productService) {
         this.productService = productService;
     }
